Implement command 'edit' functionality
